+++ modules/ascii2d.js
@@ -3,6 +3,7 @@ import Cheerio from 'cheerio';
 import CQ from './CQcode';
 import config from './config';
 import logger2 from './logger2';
+import pixivShorten from './urlShorten/pixiv';
 
 const hosts = config.ascii2dHost;
 let hostsI = 0;
@@ -74,20 +75,6 @@ function getShareText({ url, title, author, thumbnail, author_url }) {
     ${pixivShorten(url)}`;
     if (author_url) { text += `\nAuthor: ${pixivShorten(author_url)}`; }
     return text;
-}
-
-/**
- * pixiv 短链接
- *
- * @param {string} url
- * @returns
- */
-function pixivShorten(url) {
-    let pidSearch = /pixiv.+illust_id=([0-9]+)/.exec(url);
-    if (pidSearch) { return 'https://pixiv.net/i/' + pidSearch[1]; }
-    let uidSearch = /pixiv.+member\.php\?id=([0-9]+)/.exec(url);
-    if (uidSearch) { return 'https://pixiv.net/u/' + uidSearch[1]; }
-    return url;
 }
 
 export default doSearch;