+++ modules/saucenao.js
@@ -6,6 +6,7 @@ import config from './config';
 import shorten from './urlShorten/is.gd';
 import { parse } from 'url';
 import logger2 from './logger2';
+import pixivShorten from './urlShorten/pixiv';
 
 const hosts = config.saucenaoHost;
 let hostsI = 0;
@@ -23,6 +24,8 @@ const exts = {
     p: 'png',
     g: 'gif',
 };
+
+const saucenaoApiKeyAddition = config.saucenaoApiKey ? { api_key: config.saucenaoApiKey } : {};
 
 /**
  * saucenao搜索
@@ -94,10 +97,11 @@ async function doSearch(imgURL, db, debug = false) {
                 if (long_remaining < 20) warnMsg += CQ.escape(`saucenao[${hostIndex}]：注意，24h内搜图次数仅剩${long_remaining}次\n`);
                 else if (short_remaining < 5) warnMsg += CQ.escape(`saucenao[${hostIndex}]：注意，30s内搜图次数仅剩${short_remaining}次\n`);
                 //相似度
-                if (similarity < 60) {
+                if (similarity < config.picfinder.saucenaoLowAcc) {
                     lowAcc = true;
                     warnMsg += CQ.escape(`相似度[${similarity}%]过低，如果这不是你要找的图，那么可能：确实找不到此图/图为原图的局部图/图清晰度太低/搜索引擎尚未同步新图\n`);
                     if (config.picfinder.useAscii2dWhenLowAcc && (db == snDB.all || db == snDB.pixiv)) { warnMsg += '自动使用 ascii2d 进行搜索\n'; }
+                    if (config.picfinder.saucenaoHideImgWhenLowAcc) thumbnail = null;
                 }
 
                 //回复的消息
@@ -193,18 +197,6 @@ async function doSearch(imgURL, db, debug = false) {
 }
 
 /**
- * pixiv 短链接
- *
- * @param {string} url
- * @returns
- */
-function pixivShorten(url) {
-    const pidSearch = /pixiv.+illust_id=([0-9]+)/.exec(url);
-    if (pidSearch) return 'https://pixiv.net/i/' + pidSearch[1];
-    return url;
-}
-
-/**
  * 链接混淆
  *
  * @param {string} url
@@ -221,8 +213,8 @@ async function confuseURL(url) {
 
 async function getShareText({ url, title, thumbnail, author_url, source }) {
     let text = `${title}
-${CQ.img(thumbnail)}
-${await confuseURL(url)}`;
+    ${thumbnail ? CQ.img(thumbnail) : config.picfinder.replys.lowAccImgPlaceholder}
+    ${await confuseURL(url)}`;
     if (author_url) text += `\nAuthor: ${await confuseURL(author_url)}`;
     if (source) text += `\nSource: ${await confuseURL(source)}`;
     return text;
@@ -241,6 +233,7 @@ function getSearchResult(host, imgURL, db = 999) {
     else if (!/^https?:\/\//.test(host)) host = `http://${host}`;
     return get(`${host}/search.php`, {
         params: {
+            ...saucenaoApiKeyAddition,
             db: db,
             output_type: 2,
             numres: 3,