+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 ## 2020
 
+### 2-21 v2.11.7
+
+- 修复通用处理完成后未停止事件传播的问题 ([#36](https://github.com/Tsuk1ko/CQ-picfinder-robot/issues/36))
+
+### 2-18 v2.11.6
+
+- WhatAnime 使用官方提供的 API
+
+### 2-03 v2.11.5
+
+- 增加 SauceNao 低相似度值自定义配置
+- 增加“SauceNao 结果相似度过低时结果缩略图的替代文字”的配置
+- 配置项变更
+  - A `picfinder.saucenaoLowAcc`
+  - A `picfinder.replys.lowAccImgPlaceholder`
+
+### 2-01 v2.11.4
+
+- 增加“SauceNao 结果相似度过低时隐藏结果缩略图”的配置
+- 配置项变更
+  - A `picfinder.saucenaoHideImgWhenLowAcc`
+
+### 1-29 v2.11.3
+
+- 增加对`http://www.pixiv.net/(artworks|users)/[0-9]+`链接的短缩
+
 ### 1-21 v2.11.2
 
 - 增加配置项用于控制是否在 saucenao 结果低相似度或配额耗尽时使用 ascii2d